refactor: Id null 검증 로직 추가

+++ src/main/java/cart/dao/ProductDao.java
@@ -3,18 +3,20 @@ package cart.dao;
 import cart.entity.product.ProductEntity;
 import java.util.List;
 import javax.sql.DataSource;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public class ProductDao {
 
-    private final JdbcTemplate jdbcTemplate;
+    private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
     private final SimpleJdbcInsert simpleJdbcInsert;
 
     public ProductDao(final DataSource dataSource) {
-        this.jdbcTemplate = new JdbcTemplate(dataSource);
+        this.namedParameterJdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
         this.simpleJdbcInsert = new SimpleJdbcInsert(dataSource)
             .withTableName("product")
             .usingColumns("name", "image_url", "price", "description")
@@ -27,13 +29,16 @@ public class ProductDao {
     }
 
     public void delete(final Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("ID값이 null일 수 없습니다.");
+        }
         final String sql = "DELETE FROM product WHERE id = ?";
-        jdbcTemplate.update(sql, id);
+        namedParameterJdbcTemplate.getJdbcTemplate().update(sql, id);
     }
 
     public List<ProductEntity> findAll() {
         final String sql = "SELECT * FROM product";
-        return jdbcTemplate.query(sql,
+        return namedParameterJdbcTemplate.query(sql,
             (rs, rowNum) -> new ProductEntity(
                 rs.getLong("id"),
                 rs.getString("name"),
